requirements.txt
pyserial==3.5
PyQt6-sip==13.5.1
PyQt6-Qt6==6.5.1
PyQt6==6.5.1
